refactor: Update discovery detail widgets
