+++ src/widgets/discovery-detail/model/constants.ts
