+++ src/widgets/discovery-detail/ui/discovery-content.tsx
